feat(search): Rename searchUtils file
