+++ components/Common/Search/States/WithSearchResult.tsx
@@ -5,7 +5,7 @@ import { pathToBreadcrumbs } from '@/components/Common/Search/utils';
 import Link from '@/components/Link';
 import { highlighter } from '@/next.orama.mjs';
 import type { SearchDoc } from '@/types';
-import { searchHitToLinkPath } from '@/util/search';
+import { searchHitToLinkPath } from '@/util/searchUtils';
 
 import styles from './index.module.css';
