+++ util/searchUtils.ts
